requirements.txt
nest_asyncio
python-telegram-bot
keybert
transformers
